+++ src/app/events/chat/HandlePOAPGM.ts
@@ -6,7 +6,6 @@ import Log, { LogUtils } from '../../utils/Log';
 import { claimForDiscord } from '../../service/poap/ClaimPOAP';
 import ServiceUtils from '../../utils/ServiceUtils';
 import ValidationError from '../../errors/ValidationError';
-import OptInPOAP from '../../service/poap/OptInPOAP';
 
 const HandlePOAPGM = async (message: Message): Promise<void> => {
 	const content: string = message.cleanContent;
@@ -25,10 +24,6 @@ const HandlePOAPGM = async (message: Message): Promise<void> => {
 	
 	try {
 		await claimForDiscord(message.author.id.toString(), null, dmChannel);
-		await OptInPOAP(message.author, dmChannel).catch(e => {
-			Log.error(e);
-			ServiceUtils.sendOutErrorMessageForDM(dmChannel).catch(Log.error);
-		});
 		
 	} catch (e) {
 		if (e instanceof ValidationError) {