+++ src/app/service/poap/end/EndPOAP.ts
@@ -27,7 +27,7 @@ import EndTwitterFlow from './EndTwitterFlow';
 import { POAPDistributionResults } from '../../../types/poap/POAPDistributionResults';
 import channelIds from '../../constants/channelIds';
 import { POAPParticipant } from '../../../types/poap/POAPParticipant';
-import { stopTrackingUserParticipation, updateDurations } from '../../../events/tracking/HandleParticipantDuringEvent';
+import { stopTrackingUserParticipation } from '../../../events/tracking/HandleParticipantDuringEvent';
 
 export default async (guildMember: GuildMember, platform: string, ctx?: CommandContext): Promise<any> => {
 	Log.debug('attempting to end poap event');
@@ -36,8 +36,6 @@ export default async (guildMember: GuildMember, platform: string, ctx?: CommandC
 	await POAPUtils.validateUserAccess(guildMember, db);
 	
 	Log.debug('authorized to end poap event');
-	
-	await ctx?.defer(true);
 	
 	if (platform == constants.PLATFORM_TYPE_TWITTER) {
 		await EndTwitterFlow(guildMember, db, ctx);
@@ -61,30 +59,25 @@ export default async (guildMember: GuildMember, platform: string, ctx?: CommandC
 	
 	Log.debug('active poap event found');
 	
-	const isDmOn: boolean = await ServiceUtils.tryDMUser(guildMember, 'Hello! I found a poap event, let me try ending it.');
 	let channelExecution: TextChannel | null = null;
 	
-	if (!isDmOn && ctx) {
+	if (ctx) {
 		await ctx.send({ content: '⚠ Please make sure this is a private channel. I can help you distribute POAPs but anyone who has access to this channel can see the POAP links! ⚠', ephemeral: true });
-	} else if (ctx) {
-		await ctx.send({ content: 'Please check your DMs!', ephemeral: true });
 	} else if (poapSettingsDoc.channelExecutionId != channelIds.DM) {
 		if (poapSettingsDoc.channelExecutionId == null || poapSettingsDoc.channelExecutionId == '') {
 			Log.debug(`channelExecutionId missing for ${guildMember.user.tag}, ${guildMember.user.id}, skipping poap end for expired event`);
 			return;
 		}
 		channelExecution = await guildMember.guild.channels.fetch(poapSettingsDoc.channelExecutionId) as TextChannel;
-		await channelExecution.send(`Hello! Below are the participants results for ${poapSettingsDoc.event}`);
+		await channelExecution.send(`Hi <@${guildMember.user.id}>! Below are the participants results for ${poapSettingsDoc.event}`);
 	}
 	
 	const currentDateISO = dayjs().toISOString();
 	Log.debug('setting poap event to false in db');
-	const duration = dayjs().diff(poapSettingsDoc.startTime, 'm', true);
 	const updateSettingsResult: UpdateWriteOpResult = await poapSettingsDB.updateOne(poapSettingsDoc, {
 		$set: {
 			isActive: false,
 			endTime: currentDateISO,
-			duration: duration,
 		},
 	});
 
@@ -113,9 +106,7 @@ export default async (guildMember: GuildMember, platform: string, ctx?: CommandC
 	if (numberOfParticipants <= 0) {
 		Log.debug('no eligible attendees found during event');
 		const eventEndMsg = `POAP event ended. No participants found for \`${channel?.name}\` in \`${channel?.guild.name}\`.`;
-		if (isDmOn) {
-			await guildMember.send({ content: eventEndMsg });
-		} else if (ctx) {
+		if (ctx) {
 			await ctx.sendFollowUp(eventEndMsg);
 		}
 		return;
@@ -141,32 +132,23 @@ export default async (guildMember: GuildMember, platform: string, ctx?: CommandC
 		],
 	};
 	
-	if (isDmOn) {
-		embedOptions = embedOptions as MessageOptions;
-		embedOptions.files = [{ name: fileName, attachment: bufferFile }];
-		await guildMember.send(embedOptions);
-	} else if (ctx) {
+	if (ctx) {
 		embedOptions = embedOptions as MessageOptionsSlash;
 		embedOptions.file = [{ name: fileName, file: bufferFile }];
-		embedOptions.ephemeral = true;
 		await ctx.send(embedOptions);
 	} else if (channelExecution != null) {
 		embedOptions = embedOptions as MessageOptions;
 		embedOptions.files = [{ name: fileName, attachment: bufferFile }];
 		await channelExecution.send(embedOptions);
 	}
-
-	if ((guildMember.id !== guildMember.user.id) && isDmOn) {
-		return guildMember.send({ content: 'Previous event ended.' }).catch(Log.error);
-	}
 	
-	const poapLinksFile: MessageAttachment = await POAPUtils.askForPOAPLinks(guildMember, isDmOn, numberOfParticipants, ctx, channelExecution);
+	const poapLinksFile: MessageAttachment = await POAPUtils.askForPOAPLinks(guildMember, numberOfParticipants, ctx, channelExecution);
 	const listOfPOAPLinks: string[] = await POAPUtils.getListOfPoapLinks(poapLinksFile);
 	const distributionResults: POAPDistributionResults = await POAPUtils.sendOutPOAPLinks(guildMember, listOfParticipants, poapSettingsDoc.event, listOfPOAPLinks);
 	if (distributionResults.didNotSendList.length > 0) {
 		await POAPUtils.setupFailedAttendeesDelivery(guildMember, distributionResults, poapSettingsDoc.event, constants.PLATFORM_TYPE_DISCORD);
 	}
-	await POAPUtils.handleDistributionResults(isDmOn, guildMember, distributionResults, channelExecution, ctx);
+	await POAPUtils.handleDistributionResults(guildMember, distributionResults, channelExecution, ctx);
 	Log.debug('POAP end complete');
 };
 
@@ -175,10 +157,7 @@ const handleEventEndForPresentParticipants = async (
 ): Promise<void> => {
 	Log.debug('starting to handle present members for end of poap event');
 	const participantsCursor: Cursor<POAPParticipant> = await getPoapParticipantsFromDB(poapSettingsDoc.voiceChannelId, poapSettingsDoc.discordServerId);
-	const db: Db = await MongoDbUtils.connect(constants.DB_NAME_DEGEN);
-	const poapParticipantsDb: Collection<POAPParticipant> = db.collection(constants.DB_COLLECTION_POAP_PARTICIPANTS);
 	for await (const participant of participantsCursor) {
-		await updateDurations(participant, poapParticipantsDb);
 		if (participant.endTime == null || participant.endTime == '') {
 			await stopTrackingUserParticipation({ id: participant.discordUserId, tag: participant.discordUserTag }, participant.discordServerId, participant.voiceChannelId, participant);
 		}