+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 4. Update poap expiration only for dates longer than 24 hours
 5. Fix duplicate poap assignment
 6. Enhance poap claim flow
+7. Remove isDMOn from all files
+8. Add logging to DB connection
 
 ## 2.7.0-RELEASE (2022-01-18)
 
@@ -200,7 +202,7 @@
    - add workflow env keys
 3. Fix redelivery for failed participants
    - add custom start messages
-   
+
 ## 1.8.4-RELEASE (2021-11-03 - 2021-11-07)
 
 1. Introduce /coordinape form request command for BanklessDAO
@@ -251,7 +253,7 @@
 
 ## 1.7.1-RELEASE (2021-10-12)
 
-1. Update slash-create dependency 
+1. Update slash-create dependency
 
 ## 1.7.0-RELEASE (2021-10-11)